alumni-scheduler/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
Flask-WTF
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.1.2
billiard==4.2.1
bleach==6.1.0
blinker==1.6.3
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
Deprecated==1.2.18
dnspython==2.8.0
email-validator==2.1.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.2.2
flask_moment
frozenlist==1.7.0
ics==0.7.2
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.5.4
limits==5.5.0
Markdown==3.5.1
markdown-it-py==4.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.6.4
ordered-set==4.1.0
packaging==25.0
Pillow
prompt_toolkit==3.0.52
propcache==0.3.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
redis==5.0.1
requests==2.32.5
rich==13.9.4
sendgrid==6.10.0
six==1.17.0
starkbank-ecdsa==2.2.0
TatSu==5.13.1
twilio==8.10.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.17.3
WTForms==3.2.1
yarl==1.20.1
# gunicorn app:app
gunicorn